Resized sprites and created field grid
To properly layout the card slots in a grid forming a field, it was necessary to know the cards sizes inside Unity. This depends on the import settings for the sprites, then I've set a 2421px/un as import setting and resized sprites for 1920x1080px target resolution. This way the card are 57mm wide inside Unity.